Move JavaScript generator
