+++ examples/src/main/java/edu/berkeley/cs/jqf/examples/js/JavaScriptCodeGenerator.java
@@ -26,7 +26,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package edu.berkeley.cs.jqf.examples.nashorn;
+package edu.berkeley.cs.jqf.examples.js;
 
 import java.util.ArrayList;
 import java.util.Arrays;
