+++ examples/src/test/java/edu/berkeley/cs/jqf/examples/rhino/CompilerTest.java
@@ -36,7 +36,7 @@ import java.nio.charset.StandardCharsets;
 
 import com.pholser.junit.quickcheck.From;
 import edu.berkeley.cs.jqf.examples.common.AsciiStringGenerator;
-import edu.berkeley.cs.jqf.examples.nashorn.JavaScriptCodeGenerator;
+import edu.berkeley.cs.jqf.examples.js.JavaScriptCodeGenerator;
 import edu.berkeley.cs.jqf.fuzz.Fuzz;
 import edu.berkeley.cs.jqf.fuzz.JQF;
 import org.apache.commons.io.IOUtils;
